+++ core/lib/sn/CmnInterface.d.ts
@@ -113,6 +113,13 @@ export declare type HArg = {
     end_ms?: number;
     join?: boolean;
     do_rec?: boolean;
+    ':id'?: string;
+    ':path'?: string;
+    ':ln'?: number;
+    ':col_s'?: number;
+    ':col_e'?: number;
+    ':idx_tkn'?: number;
+    ':token'?: string;
 };
 export interface ITag {
     (hArg: HArg): boolean;
@@ -166,7 +173,7 @@ export interface ISysBase {
     pre(ext: string, data: string): Promise<string>;
     addHook(fnc: IFncHook): void;
     callHook: IFncHook;
-    sendDbg: IFncHook;
+    send2Dbg: IFncHook;
     copyBMFolder(from: number, to: number): void;
     eraseBMFolder(place: number): void;
 }