+++ core/lib/sn/Layer.d.ts
@@ -1,5 +1,6 @@
 import { DisplayObject, Container, Sprite, Renderer } from 'pixi.js';
 import { HArg } from './CmnInterface';
+import { IGenerateDesignCast } from './LayerMng';
 export declare class Layer {
     layname: string;
     protected name_: string;
@@ -45,6 +46,7 @@ export declare class Layer {
     snapshot(rnd: Renderer, re: () => void): void;
     snapshot_end(): void;
     cvsResize(): void;
+    drawDesignCast(_gdc: IGenerateDesignCast): void;
     dump(): string;
     static setXY(base: DisplayObject, hArg: HArg, ret: Container, isGrp?: boolean, isButton?: boolean): void;
     static setXYByPos(base: DisplayObject, pos: string, ret: DisplayObject): void;